# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication API",
    "version": "1.0.0",
    "description": "API documentation for user authentication."
  },
  "basePath": "/",
  "paths": {
...
    "/time-records/{id}": {
      "delete": {
        "summary": "Delete a time record by ID",
        "description": "Delete a time record by its unique identifier.",
        "tags": [
          "Time Records"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the time record to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Time record deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Time record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during deletion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a time record by ID",
        "description": "Update a time record by its unique identifier.",
        "tags": [
          "Time Records"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the time record to update."
          },
          {
            "in": "body",
            "name": "timeRecord",
            "description": "The updated time record object.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TimeRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time record updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Time record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during update.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
...
}